requirements.txt
djangorestframework==3.11.0
django_extensions==2.2.9
django-jsonfield==1.4.0